python/requirements.txt
requests
inquirer
pygments
pyfiglet
websocket_client
